refactor(리뷰반영): domain의 final 제거 및 article id Long타입으로 변경

+++ src/main/java/kr/codesqaud/cafe/controller/dto/ArticleDTO.java
@@ -14,11 +14,11 @@ public class ArticleDTO {
     @Pattern(regexp = "^[\\s\\S]{3,1000}$")
     private final String content;
 
-    private final Integer id;
+    private final Long id;
 
     private final String date;
 
-    public ArticleDTO(String title, String content,Integer id,String date) {
+    public ArticleDTO(String title, String content,Long id,String date) {
         this.title = title;
         this.content = content;
         this.id = id;
